refactor: 오타 수정, DrinkType관련 수정 및 cycle 오류 수정
