+++ src/entities/meetings/index.ts
@@ -1,4 +1,4 @@
 export { default as useMeetingListQuery } from './model/hooks/useMeetingListQuery';
 export type { MeetingFilters, MeetingListInfo, MeetingCardInfo } from './model/types';
-export { DrinkType, SortType } from './model/types';
+export { SortType } from './model/types';
 //
