+++ src/entities/meetings/model/types.ts
@@ -1,9 +1,4 @@
-export enum DrinkType {
-  coffee = 'COFFEE',
-  whisky = 'WHISKY',
-  wine = 'WINE',
-  end = 'END',
-}
+import { DrinkType } from '@/shared';
 
 export enum SortType {
   latest = 'LATEST',
