+++ src/entities/reviews/model/types.ts
@@ -26,14 +26,8 @@ export interface ReviewList {
   pageInfo: PageInfo;
 }
 
-export enum PageFlag {
-  MYREIVEWS,
-  MEETINGDETAIL,
-}
-
 export interface ReviewFilters {
   page: number;
   size: number;
-  pageFlag: number;
   meetingId?: number;
 }
