+++ src/entities/tasting-list/api/getTastingList.ts
@@ -2,5 +2,5 @@ import { getApi } from '@/shared/api/axiosApis';
 import { TastingList } from '@/entities/tasting-list/model/types';
 
 export default async function getTastingList(meetingId: number): Promise<TastingList | null> {
-  return getApi<TastingList>(`/api/v1/tasing-list/${meetingId}`);
+  return getApi<TastingList>(`/api/v1/tasting-list/${meetingId}`);
 }
