+++ src/shared/constants/index.ts
@@ -1,0 +1,1 @@
+export { DrinkType } from './types';
