+++ src/shared/constants/types.ts
@@ -1,0 +1,6 @@
+export enum DrinkType {
+  coffee = 'COFFEE',
+  whisky = 'WHISKY',
+  wine = 'WINE',
+  end = 'END',
+}
