+++ src/shared/lib/form/model/types.ts
@@ -73,3 +73,9 @@ export interface UseFieldValueOptions {
     | 'AddressField';
   fieldName?: string;
 }
+
+export interface LocationInfo {
+  sido: string;
+  address: string;
+  detail: string;
+}
