+++ src/shared/lib/index.ts
@@ -1,3 +1,4 @@
 export { default as useGlobalErrorStore } from './error/useGlobalErrorStore';
 export { default as useAppForm } from './form/model/hooks/useAppForm';
 export { default as compressImage } from './compressImage';
+export type { LocationInfo } from './form/model/types';
