+++ src/widgets/post-meeting/PostMeetingView.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from '@/shared/ui';
 import { useAppForm } from '@/shared/lib';
-import { DrinkType } from '@/entities/meetings/model/types';
+import { DrinkType } from '@/shared';
 import { useUploadImageMutation } from '@/features/upload-image';
 import { PostMeetingRequest } from '@/features/meetings/model/types';
 import ErrorField from '@/shared/ui/form/ErrorField';
